Rewamp ethereum structure and strategy structure
