+++ ethereum/geth/backend.go
@@ -1,0 +1,236 @@
+package geth
+
+/*
+EthereumBackend struct implements Backend
+- starts only the blockchain, evm and rpc layer and nothing else
+# interface level
+- needs to listen to transactions created over geth-rpc
+- needs to forward those transactions to the tendermint rpc
+- needs to be able to communicate with the underlying ethereum blockchain and write transactions
+- needs to forward tendermint queries to the ethereum client
+- deliverTx takes an address to deposit the transaction fee too
+- needs to create new coins from scratch when received over IBC and destroy them when sending over IBC
+- ability to define the log level
+# private
+- needs to be configurable through a go-ethereum/parity config file
+  - homestead or not
+  - gas price
+  - gas limit
+- needs to disable PoW and only validate that transactions are correct state changes
+- needs to be able to credit transactions fees to validator accounts
+- shouldn't start the networking layer
+- shouldn't start mining
+- should only expose the correct RPC interfaces
+  - syncronization message should be fired by tendermint core in a regular basis
+*/
+
+import (
+	"encoding/json"
+	"math/big"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core"
+	"github.com/ethereum/go-ethereum/core/state"
+	ethTypes "github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/eth"
+	"github.com/ethereum/go-ethereum/event"
+	"github.com/ethereum/go-ethereum/node"
+	"github.com/ethereum/go-ethereum/p2p"
+	"github.com/ethereum/go-ethereum/rpc"
+
+	abci "github.com/tendermint/abci/types"
+
+	tmClient "github.com/tendermint/tendermint/rpc/lib/client"
+)
+
+//----------------------------------------------------------------------
+
+// Backend manages the underlying ethereum state for storage and processing,
+// and maintains the connection to Tendermint for forwarding txs
+
+// Backend handles the chain database and VM
+// #stable - 0.4.0
+type Backend struct {
+	// backing ethereum structures
+	config   *eth.Config
+	ethereum *eth.Ethereum
+
+	// txBroadcastLoop subscription
+	txSub *event.TypeMuxSubscription
+
+	// pending ...
+	pending *pending
+
+	// client for forwarding txs to tendermint
+	client tmClient.HTTPClient
+}
+
+// NewBackend creates a new Backend
+// #stable - 0.4.0
+func NewBackend(ctx *node.ServiceContext, config *eth.Config,
+	client tmClient.HTTPClient) (*Backend, error) {
+	p := newPending()
+
+	// eth.New takes a ServiceContext for the EventMux, the AccountManager,
+	// and some basic functions around the DataDir.
+	ethereum, err := eth.New(ctx, config, p)
+	if err != nil {
+		return nil, err
+	}
+
+	// send special event to go-ethereum to switch homestead=true
+	currentBlock := ethereum.BlockChain().CurrentBlock()
+	ethereum.EventMux().Post(core.ChainHeadEvent{currentBlock}) // nolint: vet, errcheck
+
+	// We don't need PoW/Uncle validation
+	ethereum.BlockChain().SetValidator(NullBlockProcessor{})
+
+	ethBackend := &Backend{
+		ethereum: ethereum,
+		pending:  p,
+		client:   client,
+		config:   config,
+	}
+	return ethBackend, nil
+}
+
+// Ethereum returns the underlying the ethereum object
+// #stable
+//func (b *Backend) Ethereum() *eth.Ethereum {
+//return b.ethereum
+//}
+
+func (b *Backend) Info() abci.ResponseInfo {
+	currentBlock := b.ethereum.BlockChain().CurrentBlock()
+	height := currentBlock.Number()
+	hash := currentBlock.Hash()
+
+	// This check determines whether it is the first time ethermint gets started.
+	// If it is the first time, then we have to respond with an empty hash, since
+	// that is what tendermint expects.
+	if height.Cmp(big.NewInt(0)) == 0 {
+		return abci.ResponseInfo{
+			Data:             "ABCIEthereum",
+			LastBlockHeight:  height.Uint64(),
+			LastBlockAppHash: []byte{},
+		}
+	}
+
+	return abci.ResponseInfo{
+		Data:             "ABCIEthereum",
+		LastBlockHeight:  height.Uint64(),
+		LastBlockAppHash: hash[:],
+	}
+}
+
+func (b *Backend) SetOption(key string, value string) (log string) {
+	return ""
+}
+
+func (b *Backend) CheckTx(tx *ethTypes.Transaction) abci.Result {
+	return b.pending.checkTx(tx)
+}
+
+// DeliverTx appends a transaction to the current block
+// #stable
+func (b *Backend) DeliverTx(tx *ethTypes.Transaction) error {
+	return b.pending.deliverTx(b.ethereum.BlockChain(), b.config, b.ethereum.ApiBackend.ChainConfig(), tx)
+}
+
+// Commit finalises the current block
+// #unstable
+func (b *Backend) Commit(receiver common.Address) (common.Hash, error) {
+	return b.pending.commit(b.ethereum, receiver)
+}
+
+// ResetWork resets the current block to a fresh object
+// #unstable
+func (b *Backend) ResetWork(receiver common.Address) error {
+	work, err := b.pending.resetWork(b.ethereum.BlockChain(), receiver)
+	b.pending.work = work
+	return err
+}
+
+// UpdateHeaderWithTimeInfo uses the tendermint header to update the ethereum header
+// #unstable
+func (b *Backend) UpdateHeaderWithTimeInfo(tmHeader *abci.Header) {
+	b.pending.updateHeaderWithTimeInfo(b.ethereum.ApiBackend.ChainConfig(), tmHeader.Time, tmHeader.GetNumTxs())
+}
+
+// GasLimit returns the maximum gas per block
+// #unstable
+func (b *Backend) GasLimit() big.Int {
+	return b.pending.gasLimit()
+}
+
+//----------------------------------------------------------------------
+// Implements: node.Service
+
+// APIs returns the collection of RPC services the ethereum package offers.
+// #stable - 0.4.0
+func (b *Backend) APIs() []rpc.API {
+	apis := b.ethereum.APIs()
+	retApis := []rpc.API{}
+	for _, v := range apis {
+		if v.Namespace == "net" {
+			v.Service = NewNetRPCService(b.config.NetworkId)
+		}
+		if v.Namespace == "miner" {
+			continue
+		}
+		if _, ok := v.Service.(*eth.PublicMinerAPI); ok {
+			continue
+		}
+		retApis = append(retApis, v)
+	}
+	return retApis
+}
+
+// Start implements node.Service, starting all internal goroutines needed by the
+// Ethereum protocol implementation.
+// #stable
+func (b *Backend) Start(srvr *p2p.Server) error {
+	go b.txBroadcastLoop()
+	return nil
+}
+
+// Stop implements node.Service, terminating all internal goroutines used by the
+// Ethereum protocol.
+// #stable
+func (b *Backend) Stop() error {
+	b.txSub.Unsubscribe()
+	b.ethereum.Stop() // nolint: errcheck
+	return nil
+}
+
+// Protocols implements node.Service, returning all the currently configured
+// network protocols to start.
+// #stable
+func (b *Backend) Protocols() []p2p.Protocol {
+	return nil
+}
+
+//----------------------------------------------------------------------
+// We need a block processor that just ignores PoW and uncles and so on
+
+// NullBlockProcessor does not validate anything
+// #unstable
+type NullBlockProcessor struct{}
+
+// ValidateBody does not validate anything
+// #unstable
+func (NullBlockProcessor) ValidateBody(*ethTypes.Block) error { return nil }
+
+// ValidateState does not validate anything
+// #unstable
+func (NullBlockProcessor) ValidateState(block, parent *ethTypes.Block, state *state.StateDB, receipts ethTypes.Receipts, usedGas *big.Int) error {
+	return nil
+}
+
+// -----------------------------------------
+// format of query data
+type jsonRequest struct {
+	Method string          `json:"method"`
+	ID     json.RawMessage `json:"id,omitempty"`
+	Params []interface{}   `json:"params,omitempty"`
+}
